удалила лишний disabled вынесла в отдельный блок genre's levels
